fix(profile): bruno + fixes
